update validate back_uri
use external link

validate some function

move file

s
