+++ src/components/Announcement/helper.ts
@@ -1,12 +1,9 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
 import { useCallback } from 'react'
-import { useNavigate } from 'react-router-dom'
 import AnnouncementApi from 'services/announcement'
 
 import { AnnouncementTemplatePopup, PopupContentAnnouncement, PopupItemType } from 'components/Announcement/type'
 import { TIMES_IN_SECS } from 'constants/index'
-import { useActiveWeb3React } from 'hooks'
-import { useChangeNetwork } from 'hooks/web3/useChangeNetwork'
 import { useAppDispatch } from 'state/hooks'
 
 const LsKey = 'ack-announcements'
@@ -44,47 +41,6 @@ export const isPopupCanShow = (
 
   const isExpired = Date.now() < startAt * 1000 || Date.now() > endAt * 1000
   return !isRead && !isExpired && isRightChain && isOwn
-}
-
-/**
- * this hook to navigate to specific url
- * detect using window.open or navigate (react-router)
- * check change chain if needed
- */
-export const useNavigateToUrl = () => {
-  const navigate = useNavigate()
-  const { chainId: currentChain } = useActiveWeb3React()
-  const { changeNetwork } = useChangeNetwork()
-
-  const redirect = useCallback(
-    (actionURL: string) => {
-      if (actionURL && actionURL.startsWith('/')) {
-        navigate(actionURL)
-        return
-      }
-      const { pathname, host, search } = new URL(actionURL)
-      if (window.location.host === host) {
-        navigate(`${pathname}${search}`)
-      } else {
-        window.open(actionURL)
-      }
-    },
-    [navigate],
-  )
-
-  return useCallback(
-    (actionURL: string, chainId?: ChainId) => {
-      try {
-        if (!actionURL) return
-        if (chainId && chainId !== currentChain) {
-          changeNetwork(chainId, () => redirect(actionURL), undefined, true)
-        } else {
-          redirect(actionURL)
-        }
-      } catch (error) {}
-    },
-    [changeNetwork, currentChain, redirect],
-  )
 }
 
 export const useInvalidateTags = (reducerPath: string) => {
