+++ src/pages/NotificationCenter/GeneralAnnouncement/AnnouncementItem.tsx
@@ -7,9 +7,9 @@ import NotificationImage from 'assets/images/notification_default.png'
 import { ReactComponent as DropdownSVG } from 'assets/svg/down.svg'
 import CtaButton from 'components/Announcement/Popups/CtaButton'
 import { formatCtaName } from 'components/Announcement/Popups/DetailAnnouncementPopup'
-import { useNavigateToUrl } from 'components/Announcement/helper'
 import { Announcement } from 'components/Announcement/type'
 import { MEDIA_WIDTHS } from 'theme'
+import { useNavigateToUrl } from 'utils/redirect'
 import { escapeScriptHtml } from 'utils/string'
 import { formatTime } from 'utils/time'
